[tests] Refactor linker detection into TestRuntime.LinkAll.cs and reuse it in ResourcesTest
The BundledResources ResourcesTest needs to know whether its own assembly was
linked (in a hot-reload-compatible build resource stripping is skipped for
reloadable/non-linked user assemblies, so bundled resources remain embedded
unless the assembly was linked). TestRuntime.IsLinkAll already answers this,
but the full TestRuntime.cs can't be compiled into BundledResources because it
re-exports types (Xamarin.Utils.ApplePlatform) that clash with other referenced
test assemblies in the "link all" app.

Move the dependency-free linker-detection members (IsLinkAll, IsLinkAny and the
LinkerSentinel helper) into a small partial file TestRuntime.LinkAll.cs that can
be compiled on its own, include it alongside TestRuntime.cs in every consuming
test project, and compile only it (not the full TestRuntime.cs) into
BundledResources. ResourcesTest now uses TestRuntime.IsLinkAll instead of a
local sentinel.

Copilot-Session: c159d2ff-39c8-4cd3-bea9-882318a1d75e
