+++ tests/EmbeddedResources/dotnet/shared.csproj
@@ -20,6 +20,9 @@
 		<Compile Include="$(RootTestsDirectory)\common\TestRuntime.cs">
 			<Link>TestRuntime.cs</Link>
 		</Compile>
+		<Compile Include="$(RootTestsDirectory)\common\TestRuntime.LinkAll.cs">
+			<Link>TestRuntime.LinkAll.cs</Link>
+		</Compile>
 		<Compile Include="$(RootTestsDirectory)\..\tools\common\ApplePlatform.cs">
 			<Link>ApplePlatform.cs</Link>
 		</Compile>
