+++ tests/common/TestRuntime.cs
@@ -2,7 +2,6 @@
 #define MONOMAC
 #endif
 
-using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
 using System.Net;
@@ -1560,39 +1559,8 @@ partial class TestRuntime {
 	}
 
 	// Determine if linkall was enabled by checking if an unused class in this assembly is still here.
-	static bool? link_all;
-	[UnconditionalSuppressMessage ("Trimming", "IL2026", Justification = "This property checks whether the trimmer is enabled by checking if a type survived trimming; it's thus trimmer safe in that the any behavioral difference when the trimmer is enabled is exactly what it's looking for.")]
-	public static bool IsLinkAll {
-		get {
-			if (!link_all.HasValue)
-				link_all = typeof (TestRuntime).Assembly.GetType (typeof (TestRuntime).FullName + "+LinkerSentinel") is null;
-			return link_all.Value;
-		}
-	}
-	class LinkerSentinel { }
-
-	// Determine if any assemblies were linked by checking if a few uncommon classes in corlib are still here.
-	static bool? link_any;
-	[UnconditionalSuppressMessage ("Trimming", "IL2026", Justification = "This property checks whether the trimmer is enabled by checking if a type survived trimming; it's thus trimmer safe in that the any behavioral difference when the trimmer is enabled is exactly what it's looking for.")]
-	public static bool IsLinkAny {
-		get {
-			if (!link_any.HasValue) {
-				var uncommonTypes = new string [] {
-					"System.Action`14",
-					"System.DBNull",
-					"System.Diagnostics.Debugger",
-					"System.Func`15",
-				};
-				link_any = false;
-				foreach (var uncommonType in uncommonTypes) {
-					link_any = typeof (int).Assembly.GetType (uncommonType) is null;
-					if (link_any == true)
-						break;
-				}
-			}
-			return link_any.Value;
-		}
-	}
+	// IsLinkAll/IsLinkAny (and the LinkerSentinel helper) live in TestRuntime.LinkAll.cs, so they can
+	// be compiled on their own into assemblies that can't compile the full TestRuntime.cs.
 
 	public static bool IsOptimizeAll {
 		get {
