+++ tests/linker/link sdk/dotnet/shared.csproj
@@ -54,6 +54,9 @@
     <Compile Include="$(RootTestsDirectory)\common\TestRuntime.cs">
         <Link>TestRuntime.cs</Link>
     </Compile>
+    <Compile Include="$(RootTestsDirectory)\common\TestRuntime.LinkAll.cs">
+        <Link>TestRuntime.LinkAll.cs</Link>
+    </Compile>
     <Compile Include="$(ThisTestDirectory)\..\ILReader.cs">
       <Link>ILReader.cs</Link>
     </Compile>
