+++ tests/monotouch-test/dotnet/shared.csproj
@@ -88,6 +88,9 @@
     <Compile Include="$(RootTestsDirectory)\common\TestRuntime.cs">
       <Link>TestRuntime.cs</Link>
     </Compile>
+    <Compile Include="$(RootTestsDirectory)\common\TestRuntime.LinkAll.cs">
+      <Link>TestRuntime.LinkAll.cs</Link>
+    </Compile>
     <Compile Include="$(RootTestsDirectory)\common\TestRuntime.RunAsync.cs">
       <Link>TestRuntime.RunAsync.cs</Link>
     </Compile>
